Keep the Annict DB work list's OFFSET within range at any page number
Compute the offset as an int64 and cast the OFFSET parameter to bigint. At
100 rows per page the int32 product wrapped negative for page numbers the
handler still accepts, and PostgreSQL rejects a negative OFFSET.

[Ja]
オフセットを int64 で計算し、OFFSET パラメータを bigint にキャストする。
1 ページ 100 件では、ハンドラーが受け付ける範囲のページ番号でも int32 同士の積が
負に折り返し、PostgreSQL は負の OFFSET を拒否していた。

+++ go/internal/query/queries/works.sql
@@ -118,7 +118,7 @@ WHERE w.deleted_at IS NULL
     )
 ORDER BY w.id DESC
 LIMIT sqlc.arg('per_page')
-OFFSET sqlc.arg('page_offset');
+OFFSET sqlc.arg('page_offset')::bigint;
 
 -- name: CountDBWorks :one
 SELECT COUNT(*)
